update UI and add default PDF generation

+++ pdf-creator/Core/Models/TableItem.swift
@@ -13,10 +13,10 @@ struct TableItem {
     let name: String
     let phoneNumber: String
     let address: String
-    let id: Double
+    let id: Int
     let dateOfBirth: String
 
-    init(name: String, phoneNumber: String, id: Double, address: String, dateOfBirth: String) {
+    init(name: String, phoneNumber: String, id: Int, address: String, dateOfBirth: String) {
         self.name = name
         self.phoneNumber = phoneNumber
         self.address = address
